Implemented Cata version of Savage Defense, and lowered LotP healing to match in-game measurements
 On branch feral
 Changes to be committed:
	deleted:    sim/druid/_savage_defense.go
	modified:   sim/druid/druid.go
	modified:   sim/druid/guardian/TestGuardian.results
	new file:   sim/druid/savage_defense.go
	modified:   sim/druid/talents.go
