+++ sim/druid/guardian/TestGuardian.results
@@ -1265,7 +1265,7 @@ dps_results: {
  value: {
   dps: 3383.29649
   tps: 15843.12915
-  dtps: 469.05636
+  dtps: 200.89407
  }
 }
 dps_results: {
@@ -1315,6 +1315,6 @@ dps_results: {
  value: {
   dps: 3718.64029
   tps: 17507.64297
-  dtps: 457.97699
+  dtps: 200.20326
  }
 }
